backend/requirements.txt
pyserial==3.5
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0